mcp/pyproject.toml
[project]
name = "python-interpreter-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "mcp[cli]>=1.19.0",
    "python-dotenv>=1.1.1",
]
